Extracted ExpectedActionGroupsEvaluator from BRD's Raging Strikes to a core file.
