+++ src/parser/core/modules/ActionWindow/index.ts
@@ -3,10 +3,12 @@ export type {EvaluatedAction} from './EvaluatedAction'
 export type {AllowedGcdsOnlyOptions} from './evaluators/AllowedGcdsOnlyEvaluator'
 export {AllowedGcdsOnlyEvaluator} from './evaluators/AllowedGcdsOnlyEvaluator'
 export {ExpectedActionsEvaluator} from './evaluators/ExpectedActionsEvaluator'
+export {ExpectedActionGroupsEvaluator} from './evaluators/ExpectedActionGroupsEvaluator'
 export {calculateExpectedGcdsForTime, ExpectedGcdCountEvaluator} from './evaluators/ExpectedGcdCountEvaluator'
 export {LimitedActionsEvaluator} from './evaluators/LimitedActionsEvaluator'
 export {NotesEvaluator} from './evaluators/NotesEvaluator'
 export type {TrackedAction, TrackedActionsOptions} from './evaluators/TrackedAction'
+export type {TrackedActionGroup, TrackedActionGroupsOptions} from './evaluators/TrackedActionGroup'
 export type {EvaluationOutput, WindowEvaluator} from './evaluators/WindowEvaluator'
 
 export {ActionWindow} from './windows/ActionWindow'
